book/requirements.txt
asgiref==3.5.2
boto3==1.26.75
botocore==1.29.75
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
decorator==5.1.1
Django==3.2.14
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.2
Pillow==9.4.0
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
pytz==2022.1
razorpay==1.3.0
requests==2.28.1
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.7
urllib3==1.26.13
validators==0.20.0
win32-setctime==1.1.0
